Revert "General UX improvements."
This reverts commit 87e5f673c8f4945d41340c42ad8cd0a46e4ec652.
